+++ Extensions/dnSpy.Debugger/dnSpy.Debugger.CorDebug/Dialogs/AttachToProcess/DotNetFrameworkAttachProgramOptionsProvider.cs
@@ -54,7 +54,7 @@ namespace dnSpy.Debugger.CorDebug.Dialogs.AttachToProcess {
 						sb.EnsureCapacity((int)chBuffer);
 						hr = rtInfo.GetVersionString(sb, ref chBuffer);
 
-						yield return new DotNetFrameworkAttachProgramOptions(process.Id, sb.ToString());
+						yield return new DotNetFrameworkAttachProgramOptions((uint)process.Id, sb.ToString());
 					}
 				}
 			}
@@ -62,13 +62,13 @@ namespace dnSpy.Debugger.CorDebug.Dialogs.AttachToProcess {
 	}
 
 	sealed class DotNetFrameworkAttachProgramOptions : AttachProgramOptions {
-		public override int ProcessId { get; }
+		public override ulong ProcessId { get; }
 		public override RuntimeId RuntimeId { get; }
 		public override string RuntimeName { get; }
 
 		readonly string debuggeeVersion;
 
-		public DotNetFrameworkAttachProgramOptions(int pid, string clrVersion) {
+		public DotNetFrameworkAttachProgramOptions(ulong pid, string clrVersion) {
 			ProcessId = pid;
 			RuntimeId = new DotNetFrameworkRuntimeId(clrVersion);
 			RuntimeName = "CLR " + clrVersion;
